New listviews etc
git-svn-id: http://dcplusplus.svn.sourceforge.net/svnroot/dcplusplus/dcplusplus/trunk@236 f586c4b0-9d0d-0410-bc51-f1fe239209fc

+++ windows/StatsFrame.cpp
@@ -49,7 +49,7 @@ void StatsFrame::drawLine(CDC& dc, StatIter begin, StatIter end, CRect& rc, CRec
 	
 	StatIter i;
 	for(i = begin; i != end; ++i) {
-		if((x - i->scroll) < rc.right)
+		if((x - (int)i->scroll) < rc.right)
 			break;
 		x -= i->scroll;
 	}
@@ -213,7 +213,7 @@ void StatsFrame::UpdateLayout(BOOL /*bResizeBars*/ /* = TRUE */) {
 
 /**
  * @file
- * $Id: StatsFrame.cpp,v 1.2 2003/10/19 15:46:34 arnetheduck Exp $
+ * $Id: StatsFrame.cpp,v 1.3 2003/11/04 20:18:15 arnetheduck Exp $
  */
